falta agregar actualizacion de ataque con movimiento de todas las unidades y explosion tanque al morir. VER TAMBIEN TEMA ENERGIA Y VELOCIDADES

+++ src/UnidadesMovibles/unidad_infanteria.cpp
@@ -3,7 +3,7 @@
 #define OBJETO_MUERTO -1
 
 UnidadInfanteria::UnidadInfanteria(int rango, int velocidad, 
-float tiempo_creacion, int costo_dinero, int vida, 
+double tiempo_creacion, int costo_dinero, int vida, 
 int id, int id_duenio, int base, int altura, std::pair<int, int> 
 centro): UnidadMovible(rango, velocidad, tiempo_creacion, costo_dinero,
 vida, id, id_duenio, base, altura, centro) {}
